requirements.txt
Flask==3.0.2
Pillow==10.2.0
numpy==1.26.4
